// src/app/models/nutrient.ts
export class Nutrients {
  foodName?: string;
  quantity?: number;
  unit?: string;
  calcium?: number;
  calories?: number;
  cholesterol?: number;
  carbohydrate?: number;
  fat?: number;
  fiber?: number;
  iron?: number;
  protein?: number;
  sodium?: number;
  sugars?: number;
  vitaminD?: number;
  potassium?: number;
  weight?: number

}